+++ include/imgui/ImGui.hpp
@@ -1,20 +1,20 @@
 #ifndef TNT_IMGUI_LIB_HPP
 #define TNT_IMGUI_LIB_HPP
 
-#define IMGEN_WIDGET_ID(N) (((__LINE__ << 8) | ((N)&0xFF)) ^ ((long long)&__FILE__))
+#define IMGEN_WIDGET_ID(N) (((__LINE__ << 8) | ((N)&0xFF)) ^ ((std::size_t)&__FILE__))
 
 #define IMGUI_ID IMGEN_WIDGET_ID(1)
 
 #include <string_view>
 
 // TODO: more widgets:
-// image, checkbox, radio, text field, horizontal slider
-// selection box
-// window with Begin()/End() to replace tnt_imgui_begin()/finish()
-// vertical and horizontal slider etc...
-// TODO: better widget id generator. (maybe) by using hash(win_title + widget text)
+// image,  radio, text field, selection box
+// TODO: faster widget id generator.
 // TODO: load from config file.
+// TODO: themes
 // TODO: add support for moving and collapsing windows.
+// TODO: make sure no widget (except text) is drawed between BeginList() and EndList() calls.
+// TODO: add support for more context (multiple windows).
 
 // TODO(maybe):
 // move slider's modified value before min/max ??
@@ -103,27 +103,35 @@ bool Begin(Window *win, std::string_view name, int x, int y,
            WindowFlags flags = WindowFlags::Default) noexcept;
 void End() noexcept;
 
-bool button(Window *win, long long id, std::string_view text, int x, int y) noexcept;
+void BeginList(Window *win, bool indexed) noexcept;
+void EndList() noexcept;
 
-bool slider_int(Window *win, long long id, int x, int y, int min_, int max_,
+bool button(Window *win, std::string_view text) noexcept;
+
+bool slider_int(Window *win, std::size_t id, int x, int y, int min_, int max_,
                 int *value) noexcept;
-bool slider_float(Window *win, long long id, int x, int y, float min_, float max_,
+bool slider_float(Window *win, std::size_t id, int x, int y, float min_, float max_,
                   float *value) noexcept;
 
-bool hslider_int(Window *win, long long id, int x, int y, int min_, int max_,
+bool hslider_int(Window *win, std::size_t id, int x, int y, int min_, int max_,
                  int *value) noexcept;
-bool hslider_float(Window *win, long long id, int x, int y, float min_, float max_,
+bool hslider_float(Window *win, std::size_t id, int x, int y, float min_, float max_,
                    float *value) noexcept;
 
-bool checkbox(Window *win, long long id, int x, int y, bool *value) noexcept;
+bool checkbox(Window *win, std::string_view text, bool *value) noexcept;
 
 // WARNING: not fully functional yet.
-int menu(Window *win, long long id, int x, int y, std::string *first,
-         std::string *last) noexcept;
+int menu(Window *win, std::string_view *options, int size) noexcept;
 
 // (maybe) return 1 if *value == max_
-void progress_bar(Window *win, long long id, int x, int y, int min_, int max_,
+void progress_bar(Window *win, std::string_view text, int min_, int max_,
                   int *value) noexcept;
+
+void newline(Window *win) noexcept;
+
+void text(Window *win, std::string_view text) noexcept;
+
+void list_item(Window *win, std::string_view text) noexcept;
 
 } // namespace ImGui
 } // namespace tnt